refactor(client): extract ListMacheteClient; inline traverse-fetch helper
- Move the six methods used solely by the `list` subcommand
  (`addable_branches`, `unmanaged_branches`, `childless_managed_branches`,
  `branches_with_overridden_fork_point`, `slidable_branches`,
  `get_slidable_after`) from `MacheteClient` into a new
  `ListMacheteClient` in `git_machete/client/list.py`, and have `cli.py`
  instantiate it for `git machete list`.
- Inline the trivial one-line wrapper `remote_enabled_for_traverse_fetch`
  into its sole caller in `traverse.py`.
- Move the comment about invalid override config (non-existent/GCed
  commit etc.) back above `has_any_fork_point_override_config` where it
  belongs - the previous reorder accidentally orphaned it.
